+++ cli/tests/test_fix_command.rs
@@ -1422,7 +1422,7 @@ fn test_fix_both_sides_of_conflict() {
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
     Fixed 3 commits of 3 checked.
-    Working copy  (@) now at: mzvwutvl d4d02bf0 (conflict) (empty) (no description set)
+    Working copy  (@) now at: mzvwutvl 47e7aa32 (conflict) (empty) (no description set)
     Parent commit (@-)      : qpvuntsm 0eae0dae a | (no description set)
     Parent commit (@-)      : kkmpptxz eb61ba8d b | (no description set)
     Added 0 files, modified 1 files, removed 0 files
@@ -1442,12 +1442,12 @@ fn test_fix_both_sides_of_conflict() {
     ");
     let output = work_dir.run_jj(["file", "show", "file", "-r", "@"]);
     insta::assert_snapshot!(output, @r"
-    <<<<<<< Conflict 1 of 1
-    %%%%%%% Changes from base to side #1
+    <<<<<<< conflict 1 of 1
+    %%%%%%% qpvuntsm c8346c1c compared with zzzzzzzz 00000000
     +CONTENT A
-    +++++++ Contents of side #2
+    +++++++ kkmpptxz 7c4518cb
     CONTENT B
-    >>>>>>> Conflict 1 of 1 ends
+    >>>>>>> conflict 1 of 1 ends
     [EOF]
     ");
 }