+++ lib/src/local_working_copy.rs
@@ -66,6 +66,7 @@ use crate::backend::TreeId;
 use crate::backend::TreeValue;
 use crate::commit::Commit;
 use crate::config::ConfigGetError;
+use crate::conflict_labels::ConflictLabels;
 use crate::conflicts;
 use crate::conflicts::ConflictMarkerStyle;
 use crate::conflicts::ConflictMaterializeOptions;
@@ -976,7 +977,11 @@ impl TreeState {
                 .iter()
                 .map(|id| TreeId::new(id.clone()))
                 .collect();
-            self.tree = MergedTree::new(self.store.clone(), tree_ids_builder.build());
+            self.tree = MergedTree::new(
+                self.store.clone(),
+                tree_ids_builder.build(),
+                ConflictLabels::from_vec(proto.conflict_labels),
+            );
         }
         self.file_states =
             FileStatesMap::from_proto(proto.file_states, proto.is_file_states_sorted);
@@ -994,6 +999,7 @@ impl TreeState {
             .iter()
             .map(|id| id.to_bytes())
             .collect();
+        proto.conflict_labels = self.tree.labels().as_slice().to_owned();
         proto.file_states = self.file_states.data.clone();
         // `FileStatesMap` is guaranteed to be sorted.
         proto.is_file_states_sorted = true;
@@ -1170,7 +1176,7 @@ impl TreeState {
         });
         trace_span!("write tree").in_scope(|| -> Result<(), BackendError> {
             let new_tree = tree_builder.write_tree()?;
-            is_dirty |= new_tree.tree_ids() != self.tree.tree_ids();
+            is_dirty |= new_tree.tree_ids_and_labels() != self.tree.tree_ids_and_labels();
             self.tree = new_tree.clone();
             Ok(())
         })?;
@@ -1932,11 +1938,16 @@ impl TreeState {
         };
         let mut changed_file_states = Vec::new();
         let mut deleted_files = HashSet::new();
+        // If a conflicted file didn't change between the two trees, but the conflict
+        // labels did, we still need to re-materialize it in the working copy.
+        let include_unchanged_conflicts = old_tree.labels() != new_tree.labels();
         let mut diff_stream = old_tree
-            .diff_stream_for_file_system(new_tree, matcher)
+            .diff_stream_for_file_system(new_tree, matcher, include_unchanged_conflicts)
             .map(async |TreeDiffEntry { path, values }| match values {
                 Ok(diff) => {
-                    let result = materialize_tree_value(&self.store, &path, diff.after).await;
+                    let result =
+                        materialize_tree_value(&self.store, &path, diff.after, new_tree.labels())
+                            .await;
                     (path, result.map(|value| (diff.before, value)))
                 }
                 Err(err) => (path, Err(err)),
@@ -2024,7 +2035,8 @@ impl TreeState {
                         marker_len: Some(conflict_marker_len),
                         merge: self.store.merge_options().clone(),
                     };
-                    let contents = materialize_merge_result_to_bytes(&file.contents, &options);
+                    let contents =
+                        materialize_merge_result_to_bytes(&file.contents, &file.labels, &options);
                     let mut file_state = self
                         .write_conflict(&disk_path, &contents, file.executable.unwrap_or(false))
                         .await?;
@@ -2053,9 +2065,9 @@ impl TreeState {
         let matcher = self.sparse_matcher();
         let mut changed_file_states = Vec::new();
         let mut deleted_files = HashSet::new();
-        let mut diff_stream = self
-            .tree
-            .diff_stream_for_file_system(new_tree, matcher.as_ref());
+        let mut diff_stream =
+            self.tree
+                .diff_stream_for_file_system(new_tree, matcher.as_ref(), false);
         while let Some(TreeDiffEntry { path, values }) = diff_stream.next().await {
             let after = values?.after;
             if after.is_absent() {
@@ -2427,7 +2439,7 @@ impl LockedWorkingCopy for LockedLocalWorkingCopy {
         // continue an interrupted update if we find such a file.
         let new_tree = commit.tree();
         let tree_state = self.wc.tree_state_mut()?;
-        if tree_state.tree.tree_ids() != new_tree.tree_ids() {
+        if tree_state.tree.tree_ids_and_labels() != new_tree.tree_ids_and_labels() {
             let stats = tree_state.check_out(&new_tree)?;
             self.tree_state_dirty = true;
             Ok(stats)
@@ -2477,7 +2489,10 @@ impl LockedWorkingCopy for LockedLocalWorkingCopy {
         mut self: Box<Self>,
         operation_id: OperationId,
     ) -> Result<Box<dyn WorkingCopy>, WorkingCopyStateError> {
-        assert!(self.tree_state_dirty || self.old_tree.tree_ids() == self.wc.tree()?.tree_ids());
+        assert!(
+            self.tree_state_dirty
+                || self.old_tree.tree_ids_and_labels() == self.wc.tree()?.tree_ids_and_labels()
+        );
         if self.tree_state_dirty {
             self.wc
                 .tree_state_mut()?